requirements.txt
pandas>=1.3.0
pdfplumber>=0.7.0
PyPDF2>=3.0.0
google-generativeai>=0.3.0
pytest>=7.0.0
openpyxl>=3.0.0
tabulate>=0.9.0
google-generativeai>=0.8.0
pdfplumber>=0.10.0
pandas>=2.0.0
pytest>=7.0.0
pathlib2>=2.3.0